+++ src/main/java/com/demo/spring/boot/SpringBootArtifactDemo/exceptions/CityNotFoundException.java
@@ -2,7 +2,7 @@ package com.demo.spring.boot.SpringBootArtifactDemo.exceptions;
 
 public class CityNotFoundException extends RuntimeException {
 
-    public CityNotFoundException(Integer id) {
+    public CityNotFoundException(Long id) {
         super("La ville avec l'ID " +   id  + " n'a pas été trouvée.");
     }
 }
