+++ src/main/java/com/demo/spring/boot/SpringBootArtifactDemo/services/interfaces/CityService.java
@@ -2,6 +2,7 @@ package com.demo.spring.boot.SpringBootArtifactDemo.services.interfaces;
 
 import com.demo.spring.boot.SpringBootArtifactDemo.dto.CityDTO;
 import com.demo.spring.boot.SpringBootArtifactDemo.entities.City;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
@@ -11,13 +12,13 @@ public interface CityService {
     City createCity(City city);
 
     // Méthode pour récupérer une ville par son ID
-    City getCityById(Integer id);
+    City getCityById(Long id);
 
     // Méthode pour mettre à jour une ville existante
-    City updateCity(Integer id, City updatedCity);
+    City updateCity(Long id, City updatedCity);
 
     // Méthode pour supprimer une ville existante
-    void deleteCityById(Integer id);
+    void deleteCityById(Long id);
 
     // Méthode pour récupérer toutes les villes dans la base de données
     List<City> getAllCities();
