+++ src/main/java/com/demo/spring/boot/SpringBootArtifactDemo/services/interfaces/VehicleService.java
@@ -3,7 +3,7 @@ package com.demo.spring.boot.SpringBootArtifactDemo.services.interfaces;
 import com.demo.spring.boot.SpringBootArtifactDemo.entities.Vehicle;
 
 public interface VehicleService {
-    Vehicle getVehicleById(int id);
+    Vehicle getVehicleById(Long id);
 
     Vehicle createVehicle(Vehicle vehicle);
 }
